+++ server/src/middleware/types/decoded-jwt.ts
